+++ docs/src/lib/layout.shared.tsx
@@ -2,11 +2,11 @@ import type { BaseLayoutProps } from 'fumadocs-ui/layouts/shared'
 
 export function baseOptions(): BaseLayoutProps {
   return {
-    githubUrl: 'https://github.com/hckhanh/afetch',
+    githubUrl: 'https://github.com/hckhanh/what-the-fetch',
     nav: {
       title: (
         <span>
-          🌐<span className='ml-2'>afetch</span>
+          🌐<span className='ml-2'>what-the-fetch</span>
         </span>
       ),
       url: '/docs',